Refactor get_feature_names
