+++ mlxtend/feature_selection/sequential_feature_selector.py
@@ -21,28 +21,7 @@ from sklearn.metrics import get_scorer
 
 from ..externals.name_estimators import _name_estimators
 from ..utils.base_compostion import _BaseXComposition
-from .utilities import _calc_score, _merge_lists, _preprocess
-
-
-def _get_featurenames(subsets_dict, feature_idx, X, feature_names):
-    """
-    X is numpy.ndarray
-    """
-    if feature_names is None or len(feature_names) == 0:
-        feature_names = [str(i) for i in range(X.shape[1])]
-
-    subsets_dict_ = deepcopy(subsets_dict)
-    for key in subsets_dict_:
-        subsets_dict_[key]["feature_names"] = tuple(
-            feature_names[idx] for idx in subsets_dict[key]["feature_idx"]
-        )
-
-    if feature_idx is None:
-        feature_idx_names = None
-    else:
-        feature_idx_names = tuple(feature_names[idx] for idx in feature_idx)
-
-    return subsets_dict_, feature_idx_names
+from .utilities import _calc_score, _get_featurenames, _merge_lists, _preprocess
 
 
 class SequentialFeatureSelector(_BaseXComposition, MetaEstimatorMixin):
